+++ tests/core/speed.odin
@@ -4,3 +4,4 @@ package tests_core
 @(require) import "crypto"
 @(require) import "hash"
 @(require) import "image"
+@(require) import "math/big"
